From: Jeffrey A Law <law@hurl.cygnus.com>
  Marc Espie writes:
  > I haven't really been able to follow thru the code (and my current build
  > is not finished yet, so I can't do the check myself), but your recent patch
  >
  > to handle weak correctly does look like __attribute__((weak)) will work
  > correctly only if HANDLE_PRAGMA_WEAK is asserted.
  >
  > This is a problem, as HANDLE_PRAGMA_WEAK is a choice, not a feature (at
  > least, that's what the gcc.info documentation seems to imply for
  > HANDLE_SYSV_PRAGMA).
When HANDLE_PRAGMA_WEAK is false, the code will continue to fail as it has
failed for years.  Not perfect by any means.  Feel free to fix it.  It is
still a choice, selecting that choice is the only way to get functional
weak support right now.

What needs to happen long term is the generic weak/alias code needs to become
unconditional and only the pragma parsing of that code be conditional.