Let weak symbols take precedence over symbols read from 2nd hand libraries.
This fixes an important bug: libc now uses weak symbols a lot.
Without this patch, dynamic linking with libraries that reference libc
would mean those symbols would not be resolved, as weak symbols are good
enough for linking, but 2nd hand libraries symbols are not.

Not 100% sure this won't trigger problems later (nothing yet), but this
fixes the immediate problem.

Approved by niklas@ and millert@