Minor cleanups.
1) Don't check for the existance of /mnt/etc/hosts twice. Since we
bail out at the first check the 2nd one is superfluous.

2) Shorten and clarify the upgrade network configuration logic.

3) Always ask if manual network configuration is desired when
upgrading. Allows the upgrader to decline to use the existing network
configuration and still have an explicit opportunity to configure the
network.

4) Don't bail out if the existing network configuration fails. Instead
(see 3)) offer manual network configuration and let the user decide if
it is worthwhile to continue.

5) Refer to actual file that is missing when bailing out, rather than
forcing user to be alert enough to prepend '/mnt'.