Fix a row of bugs in the code that computes the alternate superblock locations based on the disklabel.
- unit bug: disklabel size is in sectors, but superblock
fs size is in fragments.
- The disklabel contains the maximum superblock size, not the actual
as used by newfs.
- problem is that mkfs reduces track/sector and secors/track to
make the cylindergroup fit in a fs block. In this case the values
in the disklabel do not predict what's in the superblock.

This fixes all these problems. It correctly predicts the location
of all superblocks on all my filesystems I have online (sized between
100MB and 1000GB).  Also tested with filesystems with different
block and fragment sizes.

Written this with a goto to make the diff easier to read.
ok krw@ deraadt@