emu(4) incorrectly overrides the requested recoding format, setting it to 16-bit slinear_le.
this appears to be the result of a mistake in porting the driver
from NetBSD.  when this driver was ported, NetBSD had some members
in it's 'audio_params' structure that OpenBSD doesn't have.  in
particular, NetBSD had 'hw_encoding' and 'hw_precision', which store
what the hardware is presenting.  the 'encoding' and 'precision' members
store what the audio layer is presenting (both in NetBSD and
OpenBSD) and can be different than what the hardware is presenting.
the original NetBSD code was setting 'hw_encoding' and 'hw_precision',
and in porting was changed to 'encoding' and 'precision'.

with this change the hardware is still presenting 16-bit slinear_le;
having the variable 'b16' set to '1' is enough to ensure this.
however now the audio layer will convert it to the requested format,
instead of always outputting 16-bit slinear_le.