If we find something is not aligned according to ALIGNED_POINTER(), we cannot then re-align it using ALIGN().  That is not portable since we have architectures where the modulo are quite different.  define an ALIGN_POINTER() macro in place, and use it in one spot.
This caused a NFS crash on sparc (which borrows mbufs and chains them
itself in insane ways).  I heard claudio and beck trying to diagnose
it from over the room when suddenly I knew exactly what it was.
blambert spent a few hours on it making sure that I wasn't insane.