Support unix domain sockets in nc(1) with -Uu.
Previously, using -U with -u was an error that was not documented
in the man page.  Now it will use a unix socket in datagram mode.

Bidirectional unix datagram communication requires a socket at
both ends, so in client mode (without -l), a temporary socket is
created so that responses from the server can be received.
If -s is specified with -U and -u, it specifies the location of
the temporary socket to create.

This was mostly written way back in 2007.  Since then, various
improvements implemented based on suggestions from guenther@,
tedu@, and nicm@. Man page help from nicm@ and jmc@.

Unix datagram support requires a small change to atomicio.c
in order to function correctly, this will be committed separately
shortly.

OK nicm@