to quote from the gospel of bus_dma.9:
   Synchronization operations are expressed from the perspective of the host
   RAM, e.g., a device -> memory operation is a READ and a memory -> device
   operation is a WRITE.

the status block that the isr reads is written to by the device.
the chip writes to memory, it is therefore a READ.

this also adds the preread sync when the map is set up and the postread
sync when the map is torn down for better symmetry. there are probably
more issues like this in the code, but this is a start.

discovered while discussing another diff with claudio@