Enable multiple screens
Since proper VT support is still lacking this will make VTs share a single
framebuffer in emulation mode.  As a result the display will be messed up
a bit if you run anything on VTs 1-8.  This can be mitigated by disabling
the apropriate lines in /etc/ttys.  This does allow VT switching between X
and the framebuffer console and makes suspend/resume do the right thing
when in X.