Fix critical section deferral and speed things up \o/.
It just so happens I don't know assembly very well, so the comparison
was wrong. After fixing it unrolled other bugs, ithread_run() would be
called on the isource of LIR_TIMER, which is wrong, for now fix it the
ugly way until all IPLs are out of the way.

While here also make crit_rundeferred() a non static since it doesn't
show up on ddb.