Remove splraise() and spl0().
Replaced some calls with crit_leave_all() which runs deferred
processing.

Remove the spl0() call in autoconf as we don't want interrupts that
early.

At that point not even ithreads were forked, in amd64 that is actually
a lie, since interrupts will only be enabled very late on the ioapic
via setredir().